Fixed reliability issue: Cast one of the operands of this multiplication operation to a "long".

+++ src/main/java/flowershop/sales/SalesController.java
@@ -388,7 +388,7 @@ public class SalesController {
 	) {
 		Product product = productService.getProductById(productId).get();
 		int tmpQuantity = (quantityInput == null || quantityInput == 0) ? 1 : quantityInput;
-		sellCart.addOrUpdateItem(product, -1 * tmpQuantity);
+		sellCart.addOrUpdateItem(product, -1 * (long)tmpQuantity);
 
 		double fp = salesService.calculateFullCartPrice(model, sellCart, true);
 		model.addAttribute("fullSellPrice", fp);
@@ -484,7 +484,7 @@ public class SalesController {
 
 		int tmpQuantity = (quantityInput == null || quantityInput == 0) ? 1 : quantityInput;
 
-		buyCart.addOrUpdateItem(product, -1 * tmpQuantity);
+		buyCart.addOrUpdateItem(product, -1 * (long)tmpQuantity);
 
 		double fp = salesService.calculateFullCartPrice(model, buyCart, false);
 		model.addAttribute("fullBuyPrice", fp);
